- Removed old DistributedState class (replaced by AkkaDistributedState) - Moved DistributedState classes in proper packages
