+++ client/src/main/scala/it/cwmp/client/model/AkkaDistributedState.scala
@@ -1,4 +1,4 @@
-package it.cwmp.client.model.game
+package it.cwmp.client.model
 
 import akka.actor.Actor.Receive
 import akka.actor.ActorRef
